requirements.txt
# Azure multi-agent trading system
agent-framework-core>=0.1.0
azure-identity>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
yfinance>=0.2.0
# SSE backend for Command Center dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
